Refactor Route Functions into Server Controllers
Creates server-side controllers to organize and modularize router
behavior and updates server.js to use these controllers.

This does not make any model changes. The app should work as expected
with this update. Further changes will require database schema changes.
